Refactor media strip adapter into dedicated package
